requirements.txt
tesseract
pytesseract
numpy
pillow